Rename views folder to view, fix HMR in app-forum/app-template
