+++ packages/@coorpacademy-app-forum/src/view/index.js
