+++ packages/@coorpacademy-app-template/src/view/index.js
